requirements.txt
# Core ML Libraries
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
imbalanced-learn==0.11.0

# MLOps Tools
mlflow==2.9.2

# API Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic>=2.5.0,<3.0.0
pydantic-settings==2.1.0

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==7.0.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
python-multipart==0.0.6

# Jupyter for exploration
jupyter==1.0.0
ipykernel==6.28.0
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0

# Logging
loguru==0.7.2
